feat: Extract pan and scroll functionality to reusable composable
